requirements.txt
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.10.1
six==1.16.0
